+++ src/components/ToDoSubmit.css
@@ -5,6 +5,8 @@
   width: 100%;
   margin: 1rem;
   border: grey solid 2px;
+  border-radius: 20px;
+  background-color: bisque;
 }
 
 .input-container {
